requirements.txt
Pillow >= 2.1.0
cx_Freeze == 5.0.2